refactor directories
moved stuff around to match common routing structure
