+++ BarkBook-UI/src/components/DynamicTable.jsx
